--- function
@@ -1,11 +1,0 @@
-<?php
-
-function calculationTask (array $taskList, $projectName) {
-    $countItem = 0;
-    foreach ($taskList as $key => $val){
-                if($val['category'] == $projectName){                   
-                   $countItem++;
-        }
-    }
-    return $countItem;
-}
