+++ system/init.php
@@ -1,0 +1,5 @@
+<?php
+require_once('function/function.php');
+
+date_default_timezone_set('Europe/Moscow');
+$title = "Дела в порядке";
